--- model/ex.xml
@@ -1,26 +1,0 @@
-<Game>
-<Game_state>
-	<board>
-	.b.b.b.b
-	b.b.b.b.
-	.b.b.b.b
-	........
-	........
-	w.w.w.w.
-	.w.w.w.w
-	w.w.w.w.
-	</board>
-	<who_move_last>
-	1
-	</who_move_last>
-	<he_kill>
-	0
-	</he_kill>
-	<finish_cell_last_move>
-	1 4
-	</finish_cell_last_move>
-	<move_to_draw>
-	15
-	</move_to_draw>
-	</Game_state>
-</Game>
