--- model/new.xml
@@ -1,50 +1,0 @@
-<Game>
-<Game_state>
-	<board>
-	.b.b.b.b
-	b.b.b.b.
-	.b.b.b.b
-	........
-	........
-	w.w.w.w.
-	.w.w.w.w
-	w.w.w.w.
-	</board>
-	<who_move_last>
-	2
-	</who_move_last>
-	<he_kill>
-	0
-	</he_kill>
-	<finish_cell_last_move>
-	-1 -1
-	</finish_cell_last_move>
-	<move_to_draw>
-	0
-	</move_to_draw>
-</Game_state>
-<Game_state>
-	<board>
-	.b.b.b.b
-	b.b.b.b.
-	.b.b.b.b
-	........
-	.w......
-	..w.w.w.
-	.w.w.w.w
-	w.w.w.w.
-	</board>
-	<who_move_last>
-	1
-	</who_move_last>
-	<he_kill>
-	0
-	</he_kill>
-	<finish_cell_last_move>
-	4 1
-	</finish_cell_last_move>
-	<move_to_draw>
-	0
-	</move_to_draw>
-</Game_state>
-</Game>
